+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- PDF report
+
+### Changed
+
+- report generation engine with [Venus Fugerit Doc](https://github.com/fugerit-org/fj-doc)
+
 ## [1.0.2] - 2025-11-26
 
 ### Fixed